[FE][refactor] Extract-Logic-To-Hooks
